+++ src/main/java/flab/just10minutes/member/service/MemberService.java
@@ -1,14 +1,14 @@
 package flab.just10minutes.member.service;
 
-import flab.just10minutes.member.domain.Member;
+import flab.just10minutes.member.Entity.Member;
 import flab.just10minutes.member.dto.SignUpRequestDto;
 import flab.just10minutes.member.repository.MemberDao;
+import flab.just10minutes.util.EntityTo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
 import java.util.Optional;
-import java.util.function.Consumer;
 
 @Service
 @RequiredArgsConstructor
@@ -18,7 +18,7 @@ public class MemberService{
 
     public void save(SignUpRequestDto signupDto) {
         findById(signupDto.getId()).ifPresent(m -> {throw new RuntimeException("already exist id");});
-        int insertCount = memberDao.save(SignUpRequestDto.to(signupDto));
+        int insertCount = memberDao.save(EntityTo.toEntity(signupDto));
         if (insertCount != 1) {
             throw new RuntimeException("Fail Command. Please Retry!");
         }
@@ -30,10 +30,4 @@ public class MemberService{
     }
 
 
-//    public void existMemberValidate(String id, String password) {
-//        Member member = findMemberById(id);
-//        if (!member.getPassword().equals(password)) {
-//            throw new IllegalStateException("아이디 혹은 비밀번호를 다시 확인해주세요.");
-//        }
-//    }
 }
